Adds Zoom Functionality
