+++ client/src/App.jsx
@@ -1,4 +1,3 @@
-
 import React from 'react';
 import Reviews from 'components/R&R/Reviews';
 import Questions from 'components/Q&A/questions';
